Refactoring get voters
